+++ cosmos_server/src/structure/ship/sync.rs
@@ -27,9 +27,9 @@ fn on_request_ship(
                 cosmos_encoder::serialize(&ServerReliableMessages::Ship {
                     entity: ev.entity,
                     body: NettyRigidBody::new(velocity, transform.rotation, NettyRigidBodyLocation::Absolute(*location)),
-                    width: structure.chunks_width() as u32,
-                    height: structure.chunks_height() as u32,
-                    length: structure.chunks_length() as u32,
+                    width: structure.chunks_width(),
+                    height: structure.chunks_height(),
+                    length: structure.chunks_length(),
                     chunks_needed: ChunksNeedLoaded {
                         amount_needed: structure.all_chunks_iter(false).len(),
                     },
